refactor: split themes, Ct catalog widgets, and overlay host parts (#3878)
Extract colonial/editorial theme builders, Cinzel preload, CtBrassDivider
painter, CtSlider state, province-detail host wiring, and region-map corner
helpers into library parts without changing public APIs or behavior.

Refs #3878
