+++ app/lib/features/game/flame/region_map/region_map_component_shared.dart
@@ -2,3 +2,21 @@ part of 'region_map_component.dart';
 
 // Palette tokens: `region_map_component_shared_palette.dart`.
 // Label placement: `region_map_component_shared_label_placement.dart`.
+
+class _CornerValues {
+  final bool nw;
+  final bool ne;
+  final bool sw;
+  final bool se;
+  final bool same;
+  final bool value;
+
+  _CornerValues({
+    required this.nw,
+    required this.ne,
+    required this.sw,
+    required this.se,
+    required this.same,
+    required this.value,
+  });
+}
